requirements.txt
Flask==3.0.0
cryptography==41.0.2
python-jose[cryptography]==3.3.0
requests==2.31.0
python_version > '3.6'
